Move constants in correct module
